# web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/results/detail/{id}": {
            "get": {
                "description": "Get details for a result.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Results"
                ],
                "summary": "Results detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The screenshot ID to load.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ConsoleLog": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "result_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Cookie": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "expires": {
                    "type": "string"
                },
                "http_only": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "result_id": {
                    "type": "integer"
                },
                "secure": {
                    "type": "boolean"
                },
                "session": {
                    "type": "boolean"
                },
                "size": {
                    "type": "integer"
                },
                "source_port": {
                    "type": "integer"
                },
                "source_scheme": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Header": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "result_id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.NetworkLog": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mime_type": {
                    "type": "string"
                },
                "remote_ip": {
                    "type": "string"
                },
                "request_type": {
                    "$ref": "#/definitions/models.RequestType"
                },
                "result_id": {
                    "type": "integer"
                },
                "status_code": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.RequestType": {
            "type": "integer",
            "enum": [
                0,
                0
            ],
            "x-enum-varnames": [
                "HTTP",
                "WS"
            ]
        },
        "models.Result": {
            "type": "object",
            "properties": {
                "console": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ConsoleLog"
                    }
                },
                "content_length": {
                    "type": "integer"
                },
                "cookies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Cookie"
                    }
                },
                "discovered_creds": {
                    "description": "DiscoveredCreds records the credentials (or \"anonymous\") that\nsuccessfully authenticated when the driver walked an automatic\ncredential ladder. Empty when the operator supplied creds\nexplicitly via URL or CLI flags, or when the protocol has no auth\nconcept. Currently populated by the RTSP driver.",
                    "type": "string"
                },
                "failed": {
                    "description": "Failed flag set if the result should be considered failed",
                    "type": "boolean"
                },
                "failed_reason": {
                    "type": "string"
                },
                "file_name": {
                    "description": "Name of the screenshot file",
                    "type": "string"
                },
                "final_url": {
                    "type": "string"
                },
                "headers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Header"
                    }
                },
                "html": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_pdf": {
                    "type": "boolean"
                },
                "network": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NetworkLog"
                    }
                },
                "perception_hash": {
                    "type": "string"
                },
                "perception_hash_group_id": {
                    "type": "integer"
                },
                "probed_at": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "response_reason": {
                    "type": "string"
                },
                "screenshot": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "technologies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Technology"
                    }
                },
                "title": {
                    "type": "string"
                },
                "tls": {
                    "$ref": "#/definitions/models.TLS"
                },
                "url": {
                    "type": "string"
                },
                "url_scheme": {
                    "type": "string"
                }
            }
        },
        "models.TLS": {
            "type": "object",
            "properties": {
                "cipher": {
                    "type": "string"
                },
                "encrypted_client_hello": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "issuer": {
                    "type": "string"
                },
                "key_exchange": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "resultid": {
                    "type": "integer"
                },
                "san_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TLSSanList"
                    }
                },
                "server_signature_algorithm": {
                    "type": "integer"
                },
                "subject_name": {
                    "type": "string"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_to": {
                    "type": "string"
                }
            }
        },
        "models.TLSSanList": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "tls_id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Tag": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "result_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Technology": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "result_id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}